+++ cmd/template/cluster/provider/templates/aws/cluster.go
@@ -7,7 +7,7 @@ import (
 	"github.com/giantswarm/microerror"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	capi "sigs.k8s.io/cluster-api/api/v1beta1"
+	capi "sigs.k8s.io/cluster-api/api/core/v1beta2"
 
 	"github.com/giantswarm/kubectl-gs/v5/internal/key"
 )
@@ -202,7 +202,7 @@ func newClusterCR(obj *v1alpha3.AWSCluster, c ClusterCRsConfig) *capi.Cluster {
 	clusterCR := &capi.Cluster{
 		TypeMeta: metav1.TypeMeta{
 			Kind:       "Cluster",
-			APIVersion: "cluster.x-k8s.io/v1beta1",
+			APIVersion: "cluster.x-k8s.io/v1beta2",
 		},
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      c.ClusterName,
@@ -213,11 +213,10 @@ func newClusterCR(obj *v1alpha3.AWSCluster, c ClusterCRsConfig) *capi.Cluster {
 			Labels: clusterLabels,
 		},
 		Spec: capi.ClusterSpec{
-			InfrastructureRef: &corev1.ObjectReference{
-				APIVersion: obj.APIVersion,
-				Kind:       obj.Kind,
-				Name:       obj.GetName(),
-				Namespace:  obj.GetNamespace(),
+			InfrastructureRef: capi.ContractVersionedObjectReference{
+				Kind:     obj.Kind,
+				Name:     obj.GetName(),
+				APIGroup: v1alpha3.SchemeGroupVersion.Group,
 			},
 		},
 	}