+++ cmd/template/nodepool/provider/common.go
@@ -4,10 +4,8 @@ import (
 	corev1alpha1 "github.com/giantswarm/apiextensions/v6/pkg/apis/core/v1alpha1"
 	"github.com/giantswarm/k8smetadata/pkg/annotation"
 	"github.com/giantswarm/k8smetadata/pkg/label"
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	capi "sigs.k8s.io/cluster-api/api/v1beta1"
-	capiexp "sigs.k8s.io/cluster-api/exp/api/v1beta1"
+	capi "sigs.k8s.io/cluster-api/api/core/v1beta2"
 )
 
 type NodePoolCRsConfig struct {
@@ -38,11 +36,11 @@ type NodePoolCRsConfig struct {
 	Namespace         string
 }
 
-func newcapiMachinePoolCR(config NodePoolCRsConfig, infrastructureRef *corev1.ObjectReference, bootstrapConfigRef *corev1.ObjectReference) *capiexp.MachinePool {
-	mp := &capiexp.MachinePool{
+func newcapiMachinePoolCR(config NodePoolCRsConfig, infrastructureRef capi.ContractVersionedObjectReference, bootstrapConfigRef capi.ContractVersionedObjectReference) *capi.MachinePool {
+	mp := &capi.MachinePool{
 		TypeMeta: metav1.TypeMeta{
 			Kind:       "MachinePool",
-			APIVersion: "cluster.x-k8s.io/v1beta1",
+			APIVersion: "cluster.x-k8s.io/v1beta2",
 		},
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      config.NodePoolName,
@@ -59,14 +57,14 @@ func newcapiMachinePoolCR(config NodePoolCRsConfig, infrastructureRef *corev1.Ob
 				annotation.MachinePoolName: config.Description,
 			},
 		},
-		Spec: capiexp.MachinePoolSpec{
+		Spec: capi.MachinePoolSpec{
 			ClusterName:    config.ClusterName,
 			Replicas:       toInt32Ptr(int32(config.NodesMin)), //nolint:gosec
 			FailureDomains: config.AvailabilityZones,
 			Template: capi.MachineTemplateSpec{
 				Spec: capi.MachineSpec{
 					ClusterName:       config.ClusterName,
-					InfrastructureRef: *infrastructureRef,
+					InfrastructureRef: infrastructureRef,
 					Bootstrap: capi.Bootstrap{
 						ConfigRef: bootstrapConfigRef,
 					},