+++ pkg/data/domain/nodepool/capi_md.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/giantswarm/k8smetadata/pkg/label"
 	"github.com/giantswarm/microerror"
-	capi "sigs.k8s.io/cluster-api/api/v1beta1"
+	capi "sigs.k8s.io/cluster-api/api/core/v1beta2"
 	runtimeClient "sigs.k8s.io/controller-runtime/pkg/client"
 )