+++ pkg/data/domain/nodepool/spec.go
@@ -8,8 +8,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	capaexp "sigs.k8s.io/cluster-api-provider-aws/v2/exp/api/v1beta2"
 	capzexp "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1beta1"
-	capi "sigs.k8s.io/cluster-api/api/v1beta1"
-	capiexp "sigs.k8s.io/cluster-api/exp/api/v1beta1"
+	capi "sigs.k8s.io/cluster-api/api/core/v1beta2"
 )
 
 type GetOptions struct {
@@ -31,7 +30,7 @@ type Resource interface {
 // node pool resources.
 type Nodepool struct {
 	MachineDeployment *capi.MachineDeployment
-	MachinePool       *capiexp.MachinePool
+	MachinePool       *capi.MachinePool
 
 	AWSMachineDeployment  *infrastructurev1alpha3.AWSMachineDeployment
 	AzureMachinePool      *capzexp.AzureMachinePool